requirements.txt
absl-py==1.4.0
affine==2.4.0
appdirs==1.4.4
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
blessed==1.20.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.0.12
click-plugins==1.1.1
click==8.1.3
cligj==0.7.2
cmake==3.25.0
contourpy==1.0.7
cycler==0.11.0
cython==3.0.0
decorator==5.1.1
docker-pycreds==0.4.0
executing==1.2.0
filelock==3.9.0
fonttools==4.39.2
fvcore==0.1.5.post20221221
gdal==3.6.2
gitdb==4.0.10
gitpython==3.1.31
google-auth-oauthlib==0.4.6
google-auth==2.16.2
gpustat==1.0.0
grpcio==1.51.3
huggingface-hub==0.13.2
idna==3.4
imagecodecs-lite==2022.9.26
imagecodecs==2023.3.16
imageio==2.27.0
imgaug==0.4.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
iopath==0.1.10
ipython==8.12.0
jedi==0.18.2
jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy-loader==0.2
lit==15.0.7
markdown==3.4.1
markupsafe==2.1.2
matplotlib-inline==0.1.6
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
networkx==3.0
numpy==1.23.5
nvidia-ml-py==11.495.46
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.0.0
pip==23.0.1
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydantic==1.9.2
pygments==2.14.0
pyparsing==3.0.9
pystac==1.4.0
python-dateutil==2.8.2
pytz==2022.7.1
pywavelets==1.4.1
pyyaml==6.0
radiant-mlhub==0.5.5
rasterio==1.3.6
requests-oauthlib==1.3.1
requests==2.27.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
segment-anything==1.0
sentry-sdk==1.17.0
setproctitle==1.3.2
setuptools==65.6.3
shapely==1.8.5.post1
simplejson==3.18.4
six==1.16.0
smmap==5.0.0
snuggs==1.4.7
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboard==2.12.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.15
timm==0.4.12
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.15.1+cu118
tqdm==4.64.1
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.7.1
urllib3==1.26.13
wandb==0.14.0
wcwidth==0.2.6
werkzeug==2.2.3
wheel==0.38.4
yacs==0.1.8
zipp==3.15.0